+++ tests/mocks/class-mailchimp-mock.php
@@ -19,6 +19,11 @@ class MailChimp {
 					'name'  => 'Test List',
 					'stats' => [ 'member_count' => 42 ],
 				],
+				[
+					'id'    => 'test-list-2',
+					'name'  => 'Test List 2',
+					'stats' => [ 'member_count' => 21 ],
+				],
 			],
 			'reports' => [
 				[
@@ -82,7 +87,17 @@ class MailChimp {
 		self::$database['members'][] = $contact;
 	}
 
+	/**
+	 * Can use the mock API?
+	 */
+	public static function is_api_configured() {
+		return get_option( 'newspack_mailchimp_api_key', false );
+	}
+
 	public static function get( $endpoint, $args = [] ) { // phpcs:ignore Squiz.Commenting.FunctionComment.Missing
+		if ( ! self::is_api_configured() ) {
+			return [];
+		}
 		if ( preg_match( '/lists\/.*\/merge-fields/', $endpoint ) ) {
 			return [
 				'merge_fields' => [
@@ -141,6 +156,9 @@ class MailChimp {
 	}
 
 	public static function post( $endpoint, $args = [] ) { // phpcs:ignore Squiz.Commenting.FunctionComment.Missing
+		if ( ! self::is_api_configured() ) {
+			return [];
+		}
 		if ( preg_match( '/lists\/.*\/merge-fields/', $endpoint ) ) {
 			return [
 				'status' => 200,